fix(overlay): widen blend_surface_pixel offset math to i64
The pixel offset was computed entirely in i32 before the usize cast -
the one spot in the file ignoring the i64 discipline used everywhere
else - overflowing (debug panic, release wrap) above ~536 M pixels.
Widen to i64 first; the coordinates are already bounds-checked, so the
product is exact and the cast safe at any surface size.

Also resolves the corner-band AA Low as by-design: the mapping pass
numerically refuted the claimed left/right asymmetry, and the residual
square inner corner at radius <= stroke_width is the mathematically
correct erosion of the rounded rect (same as CSS border-radius <
border-width) in a range the only caller cannot reach - it derives its
radius as scaled_radius + stroke_width. The erosion semantics and the
radius >= stroke convention are now documented on the function instead
of left for rediscovery.

Low-severity sweep items (issue #16).

+++ crates/captastic-windows/src/overlay.rs
@@ -3298,6 +3298,14 @@ fn rounded_rect_coverage(width: i32, height: i32, radius: f32, x: i32, y: i32) -
     ((inside * 255 + 32) / 64) as u8
 }
 
+/// Strokes `rect` inward by `stroke_width` with an antialiased rounded outer edge.
+///
+/// The stroke region is the true erosion of the rounded rect: the inner boundary is a
+/// concentric arc of radius `corner_radius - stroke_width`, degenerating to a square corner
+/// when the radius does not exceed the stroke - exactly like a square frame, whose corners
+/// also measure thicker diagonally. Callers wanting a constant-width stroke along the whole
+/// arc therefore pass `corner_radius >= stroke_width` (the window-tile caller derives the
+/// radius as `scaled_radius + stroke_width`, which guarantees it).
 fn draw_antialiased_rounded_outline(
     surface: &FrozenSurface,
     rect: UiRect,
@@ -3359,7 +3367,7 @@ fn blend_surface_pixel(surface: &FrozenSurface, x: i32, y: i32, foreground: [u8;
     if x < 0 || y < 0 || x >= surface.width || y >= surface.height {
         return;
     }
-    let offset = ((y * surface.width + x) * 4) as usize;
+    let offset = (i64::from(y) * i64::from(surface.width) + i64::from(x)) as usize * 4;
     // SAFETY: The checked coordinates address four writable bytes in the live off-screen DIB.
     unsafe {
         for (channel, foreground) in foreground.into_iter().enumerate() {
